Move router tier helpers to shared module
